+++ services/tracking/tracking_server/src/main/java/org/changppo/tracking/exception/common/CommonExceptionHandler.java
@@ -1,6 +1,5 @@
-package org.changppo.tracking.exception;
+package org.changppo.tracking.exception.common;
 
-import jakarta.servlet.http.HttpServletRequest;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
