+++ services/tracking/tracking_server/src/main/java/org/changppo/tracking/jwt/exception/JwtAuthenticationException.java
@@ -1,6 +1,6 @@
 package org.changppo.tracking.jwt.exception;
 
-import org.changppo.tracking.exception.ErrorCode;
+import org.changppo.tracking.exception.common.ErrorCode;
 
 public class JwtAuthenticationException extends JwtBusinessException{
     public JwtAuthenticationException() {
