+++ services/tracking/tracking_server/src/main/java/org/changppo/tracking/jwt/exception/JwtTokenExpiredException.java
@@ -1,6 +1,6 @@
 package org.changppo.tracking.jwt.exception;
 
-import org.changppo.tracking.exception.ErrorCode;
+import org.changppo.tracking.exception.common.ErrorCode;
 
 public class JwtTokenExpiredException extends JwtBusinessException{
     public JwtTokenExpiredException() {
